refactor: remove services from store
