+++ src/stores/useStore.ts
@@ -5,18 +5,13 @@ import type {
   Theme,
   Tour,
   TourStep,
-  UsetifulResponse,
   UsetigulTag,
 } from '../types';
 import { type LayoutChangeEvent } from 'react-native';
 import { createJSONStorage, persist } from 'zustand/middleware';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import { THEME_DEFAULT } from '../constants';
-
-const BaseURl = 'https://www.usetiful.com';
-// const BaseURl = 'https://admin:admin123@dev.usetiful.com';
-
-const END_POINT = '/api-space/data.json?lang=en&app=mobile';
+import { fetchDataJson, fetchProgressor } from '../services/api';
 
 interface StoreState {
   token: string | undefined;
@@ -190,94 +185,3 @@ export const useStore = create(
     }
   )
 );
-
-const fetchDataJson = async (token: string): Promise<Tour[]> => {
-  const reqUrl = `${BaseURl}${END_POINT}`;
-  try {
-    const response = await fetch(reqUrl, {
-      method: 'GET',
-      headers: {
-        'X-Auth-Token': token,
-        'X-Requested-With': 'XMLHttpRequest',
-        'Content-Type': 'application/json; charset=utf-8',
-      },
-    });
-
-    if (!response.ok) {
-      console.error(`USETIFUL HTTP ERROR - status: ${response.status}`);
-      return [];
-    }
-
-    const res: UsetifulResponse = await response.json();
-
-    console.log(`
-      =============================================
-      =============================================
-      ============== USETIFUL =====================
-      ================= IS ========================
-      ============== LOADED =======================
-      =============================================
-      =============================================`);
-
-    return res.tours;
-  } catch (error: any) {
-    console.error('=====error====>', error.message);
-    return [];
-  }
-};
-
-const fetchProgressor = async (token: string, userId: string) => {
-  const url = 'https://progressor.usetiful.com/api/get';
-  const headers = {
-    'Content-Type': 'application/json',
-    'X-Requested-With': 'XMLHttpRequest',
-  };
-  const body = JSON.stringify({
-    userId,
-    accountToken: token,
-  });
-
-  try {
-    const response = await fetch(url, {
-      method: 'POST',
-      headers: headers,
-      body: body,
-    });
-
-    if (!response.ok) {
-      console.error(
-        '=======Error=====>',
-        new Error(`Usetiful: connection error ${response.status}`)
-      );
-      return null;
-    }
-    const result = JSON.parse(await response.json());
-
-    let tours = [];
-    if (result.tours) {
-      try {
-        tours = JSON.parse(result.tours);
-      } catch (error) {
-        console.warn("Warning: 'tours' key is not a valid JSON string.");
-      }
-    } else {
-      console.warn("Warning: 'tours' key not found in response.");
-    }
-
-    const autoSegment = result.autoSegment;
-    const storedAt = result.storedAt;
-    const customSegments = result.customSegments;
-    const uf_completed = result.uf_completed;
-
-    return {
-      tours,
-      autoSegment,
-      customSegments,
-      storedAt,
-      uf_completed,
-    } as ProgressorData;
-  } catch (error: any) {
-    console.error('=======Error=====>', error.message);
-    return null;
-  }
-};
